+++ .github/workflows/build-pr.yml
@@ -14,3 +14,5 @@ jobs:
           java-version: '17'
       - name: Build project with Maven
         run: mvn -B -ntp -Dstyle.color=always install
+      - name: Coveralls GitHub Action
+        uses: coverallsapp/github-action@v2.2.3
