+++ cli/src/main/java/com/devonfw/tools/ide/context/AbstractIdeContext.java
@@ -667,8 +667,9 @@ public abstract class AbstractIdeContext implements IdeContext {
     assert (options.length >= 2);
     interaction(question);
     Map<String, O> mapping = new HashMap<>(options.length);
-    int i = 1;
+    int i = 0;
     for (O option : options) {
+      i++;
       String key = "" + option;
       addMapping(mapping, key, option);
       String numericKey = Integer.toString(i);
